+++ src/main/java/com/Dao/RedPacketBySqlDao.java
@@ -13,10 +13,10 @@ import java.util.List;
 @Mapper
 public  interface RedPacketBySqlDao {
     //从发布红包人减去金额
-    int deductRedPacketMoney(@Param("userId") int userId, @Param("money") double money);
+    int deductRedPacketMoney(@Param("userId") long userId, @Param("money") double money);
 
     //增加抢到红包人的账户
-    int addMoneyToUser(@Param("userId") int userId, @Param("money") double money);
+    int addMoneyToUser(@Param("userId") long userId, @Param("money") double money);
 
     //将红包细节持久化到Mysql
     int insertRedPacketDetail(@Param("list") List<UserRedPacket> partlist,double moneysum,long redpacketid);
